symplique-cosmos-optimization/requirements-dev.txt
azure-functions
azure-cosmos
azure-storage-blob
pytest
pytest-cov
python-dotenv